requirements.txt
Flask
flask-cors
openai
python-dotenv
Pillow
requests
psycopg2-binary
gunicorn